prost-build: Allow generation without FileDescriptor.source_info
Current prost assumes the presence of
`FileDescriptorProto.source_code_info`, this is used to generate and
attach comments from the protobuf files.

As per the specification this field is optional and is not always set in
every protobuf toolchain.

```proto
  // This field contains optional information about the original source code.
  // You may safely remove this entire field without harming runtime
  // functionality of the descriptors -- the information is needed only by
  // development tools.
  optional SourceCodeInfo source_code_info = 9;
```

This appears in complex rules like those present with bazel. The
sequence here is as follows:

1) The descriptor set is generated initially with protoc

```sh
  protoc \
    '--descriptor_set_out=…/xyz_proto-descriptor-set.proto.bin' \
    '-Iprotos/barcode.proto=…/xyz.proto' \
    --direct_dependencies …/xyz.proto \
    '--direct_dependencies_violation_msg=%s is imported but not used …' \
    protos/barcode.proto
```

2) This is then given to the generator plugin

```sh
  protoc \
    --descriptor_set_in=…/xyz_proto-descriptor-set.proto.bin \
    --plugin=protoc-gen-rust_plugin=…/protoc-gen-prost \
    --rust_plugin_out=…/_rpg_premerge_test_pb \
    protos/barcode.proto
```

Notice that the following flag (and, naturally, its effects) was not given to
protoc

```sh
  --include_source_info       When using --descriptor_set_out, do not strip
                              SourceCodeInfo from the FileDescriptorProto.
                              This results in vastly larger descriptors that
                              include information about the original
                              location of each decl in the source file as
                              well as surrounding comments.
```

This results in protobuf compilation failing as we attempt to unwrap the
`Option<SourceCodeInfo>`.

As such the code is changed to account for this. While it is possible to
specify the `--include_source_info` flag (and IMHO this is probably a good
idea) it is not gaurenteed to be the default behaviour of preexisting
toolchains.

With the behaviour of treating the optional sourcecode information as per the
protobuf definition this plugin will behave in the same fashion as most other
known protoc plugins.